+++ shell/src/main/java/alluxio/cli/bundler/command/CollectAlluxioInfoCommand.java
@@ -24,7 +24,7 @@ import java.nio.file.Paths;
  * Command to run a set of Alluxio commands.
  * Collects information about the Alluxio cluster.
  * */
-public class CollectAlluxioInfoCommand extends CollectEnvCommand {
+public class CollectAlluxioInfoCommand extends ExecuteShellCollectInfoCommand {
   public static final String COMMAND_NAME = "collectAlluxioInfo";
   private static final Logger LOG = LoggerFactory.getLogger(CollectAlluxioInfoCommand.class);
 
@@ -60,7 +60,8 @@ public class CollectAlluxioInfoCommand extends CollectEnvCommand {
     }
   }
 
-  private void registerCommands() {
+  @Override
+  protected void registerCommands() {
     // TODO(jiacheng): a command to find lost blocks?
     registerCommand("getConf",
             new AlluxioCommand(mAlluxioPath, "getConf --master --source"), null);
@@ -83,87 +84,6 @@ public class CollectAlluxioInfoCommand extends CollectEnvCommand {
   public String getCommandName() {
     return COMMAND_NAME;
   }
-
-//  @Override
-//  public int run(CommandLine cl) throws AlluxioException, IOException {
-//    int ret = 0;
-//
-//    // Determine the working dir path
-//    mWorkingDirPath = getWorkingDirectory(cl);
-//
-//    StringWriter output = new StringWriter();
-//    for (AlluxioCommand cmd : getAlluxioCommands()) {
-//      CommandReturn cr;
-//      String crStr = "";
-//      int cmdExitCode = 0;
-//      boolean cmdCompleted = false;
-//      try {
-//        cr = cmd.runWithOutput();
-//        cmdExitCode = cr.getExitCode();
-//        if (cr.getExitCode() != 0) {
-//          crStr = String.format("Command %s failed: %s", cmd, cr.getFormattedOutput());
-//          LOG.warn(crStr);
-//        } else {
-//          // Command completed
-//          crStr = String.format("Command %s succeeded %s", cmd, cr.getFormattedOutput());
-//          LOG.info(crStr);
-//          cmdCompleted = true;
-//        }
-//      } catch (IOException e) {
-//        crStr = String.format("Command %s failed with exception %s", cmd, e.getMessage());
-//        LOG.warn(crStr);
-//        LOG.debug("%s", e);
-//      }
-//      output.write(crStr);
-//
-//      if (!cmdCompleted) {
-//        if (!cmd.hasAlternativeCommand()) {
-//          String noAltMsg = String.format("No alternative command for command %s", cmd);
-//          output.write(noAltMsg);
-//          continue;
-//        } else {
-//          // Try alternative command
-//          ShellCommand alternativeCmd = cmd.getAlternativeCommand();
-//          String tryAgainMsg = String.format("Try alternative command %s", alternativeCmd);
-//          output.write(tryAgainMsg);
-//          LOG.info(tryAgainMsg);
-//
-//          String tryAgainRes = "";
-//          try {
-//            CommandReturn tryAgain = alternativeCmd.runWithOutput();
-//            if (tryAgain.getExitCode() != 0) {
-//              cmdExitCode = tryAgain.getExitCode();
-//              tryAgainRes = String.format("Alternative command %s failed: %s", alternativeCmd,
-//                      tryAgain.getFormattedOutput());
-//              LOG.warn(tryAgainRes);
-//            } else {
-//              cmdExitCode = 0;
-//              tryAgainRes = String.format("Alternative command %s succeeded: %s", alternativeCmd,
-//                      tryAgain.getFormattedOutput());
-//              LOG.info(tryAgainRes);
-//            }
-//          } catch (IOException e) {
-//            tryAgainRes = String.format("Alternative command %s failed with exception: %s",
-//                    alternativeCmd, e.getMessage());
-//            LOG.warn(tryAgainRes);
-//            LOG.debug("%s", e);
-//            ret = 1;
-//          }
-//          output.write(tryAgainRes);
-//        }
-//      }
-//      // Keep only the larger return code
-//      if (cmdExitCode > ret) {
-//        ret = cmdExitCode;
-//      }
-//    }
-//
-//    // TODO(jiacheng): phase 2 consider outputting partial results in a finally block
-//    File outputFile = generateOutputFile(mWorkingDirPath, OUTPUT_FILE_NAME);
-//    FileUtils.writeStringToFile(outputFile, output.toString());
-//
-//    return ret;
-//  }
 
   @Override
   public String getUsage() {
