+++ shell/src/main/java/alluxio/cli/bundler/command/CollectEnvCommand.java
@@ -12,27 +12,19 @@
 package alluxio.cli.bundler.command;
 
 import alluxio.client.file.FileSystemContext;
-import alluxio.collections.Pair;
 import alluxio.conf.PropertyKey;
-import alluxio.exception.AlluxioException;
-import alluxio.shell.CommandReturn;
 import alluxio.shell.ShellCommand;
 
-import org.apache.commons.cli.CommandLine;
-import org.apache.commons.io.FileUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.File;
-import java.io.IOException;
-import java.io.StringWriter;
 import java.util.HashMap;
 import java.util.Map;
 
 /**
- * Command to run a set of shell commands to get system information.
+ * Command to run a set of bash commands to get system information.
  * */
-public class CollectEnvCommand extends AbstractInfoCollectorCommand {
+public class CollectEnvCommand extends ExecuteShellCollectInfoCommand {
   public static final String COMMAND_NAME = "collectEnv";
   private static final Logger LOG = LoggerFactory.getLogger(CollectEnvCommand.class);
 
@@ -51,7 +43,8 @@ public class CollectEnvCommand extends AbstractInfoCollectorCommand {
     registerCommands();
   }
 
-  private void registerCommands() {
+  @Override
+  protected void registerCommands() {
     registerCommand("ps", new ShellCommand(new String[]{"ps", "-ef", "|grep alluxio*"}), null);
     registerCommand("env", new ShellCommand(new String[]{"env"}), null);
     registerCommand("top", new ShellCommand(new String[]{"atop", "-b", "-n", "1"}),
@@ -72,13 +65,6 @@ public class CollectEnvCommand extends AbstractInfoCollectorCommand {
     registerCommand("dstat", new ShellCommand(new String[]{"dstat", "-cdgilmnprsty"}), null);
   }
 
-  protected void registerCommand(String name, ShellCommand cmd, ShellCommand alternativeCmd) {
-    mCommands.put(name, cmd);
-    if (alternativeCmd != null) {
-      mCommandsAlt.put(name, alternativeCmd);
-    }
-  }
-
   @Override
   public String getCommandName() {
     return COMMAND_NAME;
@@ -87,79 +73,6 @@ public class CollectEnvCommand extends AbstractInfoCollectorCommand {
   @Override
   public boolean hasSubCommand() {
     return false;
-  }
-
-  protected Pair<Integer, String> runAndFormatOutput(ShellCommand cmd) {
-    String crStr;
-    int cmdExitCode;
-    try {
-      CommandReturn cr = cmd.runWithOutput();
-      cmdExitCode = cr.getExitCode();
-      if (cr.getExitCode() != 0) {
-        crStr = String.format("Command %s failed: %s", cmd, cr.getFormattedOutput());
-        LOG.warn(crStr);
-      } else {
-        // Command succeeded
-        crStr = String.format("Command %s succeeded %s", cmd, cr.getFormattedOutput());
-        LOG.info(crStr);
-      }
-    } catch (IOException e) {
-      cmdExitCode = 1;
-      crStr = String.format("Command %s failed with exception %s", cmd, e.getMessage());
-      LOG.warn(crStr);
-      LOG.debug("%s", e);
-    }
-    return new Pair<>(cmdExitCode, crStr);
-  }
-
-  @Override
-  public int run(CommandLine cl) throws AlluxioException, IOException {
-    int ret = 0;
-
-    // Determine the working dir path
-    mWorkingDirPath = getWorkingDirectory(cl);
-
-    // Output buffer stream
-    StringWriter outputBuffer = new StringWriter();
-
-    for (String cmdName : mCommands.keySet()) {
-      int cmdExitCode = 0;
-
-      ShellCommand cmd = mCommands.get(cmdName);
-      Pair<Integer, String> cmdOutput = runAndFormatOutput(cmd);
-      outputBuffer.write(cmdOutput.getSecond());
-
-      if (cmdOutput.getFirst() == 0) {
-        continue;
-      } else {
-        cmdExitCode = cmdOutput.getFirst();
-      }
-
-      // if there is a backup option, try it first
-      if (mCommandsAlt.containsKey(cmdName)) {
-        ShellCommand betterCmd = mCommandsAlt.get(cmdName);
-        Pair<Integer, String> cmdAltOutput = runAndFormatOutput(betterCmd);
-        outputBuffer.write(cmdAltOutput.getSecond());
-        // If the backup option succeeded, count as successful
-        if (cmdAltOutput.getFirst() == 0) {
-          cmdExitCode = 0;
-        }
-      }
-
-      // Keep only the larger return code
-      if (cmdExitCode != 0 && ret == 0) {
-        ret = cmdExitCode;
-      }
-    }
-
-    // output the logs
-    File outputFile = generateOutputFile(mWorkingDirPath,
-            String.format("%s.txt", getCommandName()));
-    LOG.info(String.format("Finished all commands. Writing to output file %s",
-            outputFile.getAbsolutePath()));
-    FileUtils.writeStringToFile(outputFile, outputBuffer.toString());
-
-    return ret;
   }
 
   @Override
