+++ shell/src/test/java/alluxio/cli/bundler/CollectInfoTest.java
@@ -13,7 +13,7 @@ package alluxio.cli.bundler;
 
 import static org.junit.Assert.assertEquals;
 
-import alluxio.cli.bundler.command.AbstractInfoCollectorCommand;
+import alluxio.cli.bundler.command.AbstractCollectInfoCommand;
 import alluxio.conf.InstancedConfiguration;
 import alluxio.cli.Command;
 import alluxio.util.ConfigurationUtils;
@@ -30,7 +30,7 @@ public class CollectInfoTest {
 
   private int getNumberOfCommands() {
     Reflections reflections =
-            new Reflections(AbstractInfoCollectorCommand.class.getPackage().getName());
+            new Reflections(AbstractCollectInfoCommand.class.getPackage().getName());
     int cnt = 0;
     for (Class<? extends Command> cls : reflections.getSubTypesOf(Command.class)) {
       if (!Modifier.isAbstract(cls.getModifiers())) {
